third_party/WebKit/LayoutTests/fast/events/wheel/wheelevent-in-text-node.html

<!DOCTYPE html>
<html>
<head>
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
  <script src="../../../resources/gesture-util.js"></script>
</head>
<script>
  setup({ explicit_done: true });
  function test() {
    var div = document.querySelector('div');
    var x = div.offsetLeft + 5;
    var y = div.offsetTop + 5;
    mouseMoveTo(x, y)
    .then(smoothScroll(120, x, y, GestureSourceType.MOUSE_INPUT, 'down'));
  }

  function wheelHandler(e) {
    window.theEvent = e;
    assert_equals(theEvent.target.nodeName, 'DIV');
    done();
  }

  window.onload = function () {
    var div = document.querySelector('div');
    div.addEventListener('mousewheel', wheelHandler);
    test();
  };
</script>
<body>
    <div>This is a div containing text. Wheel events originating on the text
    node should target the div.</div>
</body>
</html>
